+++ bootstrap/app.php
@@ -15,9 +15,10 @@ return Application::configure(basePath: dirname(__DIR__))
         $middleware->validateCsrfTokens(except: [
             'api/*',
         ]);
-        // Check if the application has been installed
+        // Check if the application has been installed & ensure web authentication
         $middleware->web(append: [
             \App\Http\Middleware\CheckInstalled::class,
+            \App\Http\Middleware\CheckWebAuth::class,
         ]);
     })
     ->withExceptions(function (Exceptions $exceptions): void {